Refactor RoutePlan, TableRoutePlan
